+++ packages/core/src/index.ts
@@ -8,6 +8,7 @@ export * from "./yahoo.js";
 export * from "./quoteProviders.js";
 export * from "./income.js";
 export * from "./drpCheck.js";
+export * from "./reconcile.js";
 export * from "./parse/index.js";
 export { parseSharesightPaste } from "./parse/sharesightPaste.js";